Move broadcast functions to list of functions for manipulating arrays
